requirements.txt
whitenoise==5.0.1
django-heroku==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
psycopg2==2.8.5
sqlparse==0.3.0
gunicorn==20.0.4
django-filter==2.2.0
django-multiselectfield==0.1.12
djangorestframework==3.11.0
